fix(Backend): :zap: fixing sequelize type
fixing sequelize type due database size and performance

+++ packages/backend/database/migrations/2-20230109202241-create-users.js
@@ -37,7 +37,7 @@ module.exports = {
       },
       emailChecked: {
         defaultValue: 0,
-        type: Sequelize.INTEGER,
+        type: Sequelize.TINYINT,
       },
       createdAt: {
         allowNull: false,
@@ -49,7 +49,7 @@ module.exports = {
       },
       cashforceAdm: {
         defaultValue: 0,
-        type: Sequelize.INTEGER,
+        type: Sequelize.TINYINT,
       },
     }, {
       initialAutoIncrement: 122,
