+++ packages/backend/database/migrations/8-20230109142245-create-orderportions.js
@@ -25,7 +25,7 @@ module.exports = {
       availableToMarket: {
         allowNull: false,
         defaultValue: 1,
-        type: Sequelize.INTEGER,
+        type: Sequelize.TINYINT,
       },
       createdAt: {
         allowNull: false,
